Task 1 submission

+++ 0x15-file_io/test/task1/main.h
@@ -11,5 +11,5 @@ typedef unsigned long int _ulint;
 ssize_t read_textfile(const char *filename, size_t letters);
 int create_file(const char *filename, char *text_content);
 int append_text_to_file(const char *filename, char *text_content);
-int _strlen(char *str);
+size_t _strlen(char *str);
 #endif
